+++ extensions/lastfm/CHANGELOG.md
@@ -1,10 +1,10 @@
-Wifi Password Reveal Changelog
+Last.fm Changelog
 
 [Fixes and New Features] - {PR_MERGE_DATE}
 
 - Grid view added
 - Time period dropdown added
-- Unknow new error code adedd
+- Unknown new error code added
 - New custom fetchLastFmData hook added
 
-[Initial Version] - 2021-01-14
+[Initial Version] - 2022-01-14